requirements.txt
streamlit>=1.29.0
openai>=1.30.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
opencv-python>=4.8.0
transformers>=4.30.0
datasets>=2.10.0
joblib>=1.3.0
python-dotenv>=1.0.0
